Use correct credentials when calling external clients
