+++ src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/client/findAndReferInterventionApi/model/FindAndReferReferralDetails.kt
@@ -1,11 +1,11 @@
 package uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.client.findAndReferInterventionApi.model
 
-import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.InterventionType
-import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.PersonReferenceType
 import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.ReferralEntity
 import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.ReferralStatusHistoryEntity
-import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.SettingType
-import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.SourcedFromReferenceType
+import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type.InterventionType
+import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type.PersonReferenceType
+import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type.SettingType
+import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type.SourcedFromReferenceType
 import java.util.UUID
 
 data class FindAndReferReferralDetails(
@@ -17,6 +17,7 @@ data class FindAndReferReferralDetails(
   val setting: SettingType,
   val sourcedFromReference: String,
   val sourcedFromReferenceType: SourcedFromReferenceType,
+  val eventNumber: String,
 )
 
 fun FindAndReferReferralDetails.toReferralEntity(statusHistories: MutableList<ReferralStatusHistoryEntity>) = ReferralEntity(
@@ -27,4 +28,5 @@ fun FindAndReferReferralDetails.toReferralEntity(statusHistories: MutableList<Re
   personName = "UNKNOWN",
   statusHistories = statusHistories,
   eventNumber = sourcedFromReference,
+  eventId = eventNumber,
 )
