+++ src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/entity/type/InterventionType.kt
@@ -1,0 +1,8 @@
+package uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type
+
+enum class InterventionType {
+  SI,
+  ACP,
+  CRS,
+  TOOLKITS,
+}
