+++ src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/entity/type/PersonReferenceType.kt
@@ -1,0 +1,6 @@
+package uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type
+
+enum class PersonReferenceType {
+  CRN,
+  NOMS,
+}
