+++ src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/entity/type/SettingType.kt
@@ -1,0 +1,8 @@
+package uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type
+
+enum class SettingType {
+  COMMUNITY,
+  CUSTODY,
+  REMAND,
+  PRE_RELEASE,
+}
