+++ src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/entity/type/SourcedFromReferenceType.kt
@@ -1,0 +1,6 @@
+package uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.entity.type
+
+enum class SourcedFromReferenceType {
+  LICENCE_CONDITION,
+  REQUIREMENT,
+}
