+++ src/main/resources/application-local.yml
@@ -17,7 +17,7 @@ spring:
           manage-and-deliver-api-client:
             provider: hmpps-auth
             client-id: hmpps-accredited-programmes-manage-and-deliver-api-client-1
-            client-secret: clientsecret
+            client-secret: ${API_CLIENT_SECRET:clientsecret}
             authorization-grant-type: client_credentials
         provider:
           hmpps-auth:
